refactor: revert class split
